Create a Log And Registration
PR Description : In order to see Login and Registration page run following command in flutter *note( directory should in the front end)  "flutter run". This will show the UI of the login and Registration Page and you'll be able to navigate the two pages by click on the button of the page the "Registration Link Or the Login Now" link.